Refactor: move client API to `application` namespace.
