+++ carp.clients.core/src/commonMain/kotlin/dk/cachet/carp/clients/application/ClientManager.kt
@@ -1,10 +1,15 @@
-package dk.cachet.carp.clients.domain
+package dk.cachet.carp.clients.application
 
+import dk.cachet.carp.clients.domain.ClientRepository
+import dk.cachet.carp.clients.domain.DeviceRegistrationStatus
 import dk.cachet.carp.clients.domain.data.ConnectedDeviceDataCollector
 import dk.cachet.carp.clients.domain.data.DataListener
 import dk.cachet.carp.clients.domain.data.DeviceDataCollector
 import dk.cachet.carp.clients.domain.data.DeviceDataCollectorFactory
+import dk.cachet.carp.clients.domain.study.Study
 import dk.cachet.carp.clients.domain.study.StudyDeploymentProxy
+import dk.cachet.carp.clients.application.study.StudyId
+import dk.cachet.carp.clients.application.study.StudyStatus
 import dk.cachet.carp.common.application.UUID
 import dk.cachet.carp.common.application.devices.DeviceRegistration
 import dk.cachet.carp.common.application.devices.DeviceRegistrationBuilder
