+++ carp.clients.core/src/commonMain/kotlin/dk/cachet/carp/clients/infrastructure/InMemoryClientRepository.kt
@@ -1,8 +1,8 @@
 package dk.cachet.carp.clients.infrastructure
 
 import dk.cachet.carp.clients.domain.ClientRepository
-import dk.cachet.carp.clients.domain.Study
-import dk.cachet.carp.clients.domain.StudySnapshot
+import dk.cachet.carp.clients.domain.study.Study
+import dk.cachet.carp.clients.domain.study.StudySnapshot
 import dk.cachet.carp.common.application.UUID
 import dk.cachet.carp.common.application.devices.DeviceRegistration
