+++ carp.clients.core/src/commonTest/kotlin/dk/cachet/carp/clients/ClientCodeSamples.kt
@@ -1,7 +1,7 @@
 package dk.cachet.carp.clients
 
 import dk.cachet.carp.clients.domain.SmartphoneClient
-import dk.cachet.carp.clients.domain.StudyStatus
+import dk.cachet.carp.clients.application.study.StudyStatus
 import dk.cachet.carp.clients.domain.createDataCollectorFactory
 import dk.cachet.carp.clients.domain.createParticipantInvitation
 import dk.cachet.carp.clients.domain.data.DataListener
