+++ carp.clients.core/src/commonTest/kotlin/dk/cachet/carp/clients/domain/study/StudyDeploymentProxyTest.kt
@@ -1,5 +1,6 @@
 package dk.cachet.carp.clients.domain.study
 
+import dk.cachet.carp.clients.application.study.StudyStatus
 import dk.cachet.carp.clients.domain.connectedDevice
 import dk.cachet.carp.clients.domain.createDataListener
 import dk.cachet.carp.clients.domain.createDependentSmartphoneStudy
@@ -9,8 +10,6 @@ import dk.cachet.carp.clients.domain.createStudyDeployment
 import dk.cachet.carp.clients.domain.deviceSmartphoneDependsOn
 import dk.cachet.carp.clients.domain.smartphone
 import dk.cachet.carp.clients.domain.DeviceRegistrationStatus
-import dk.cachet.carp.clients.domain.Study
-import dk.cachet.carp.clients.domain.StudyStatus
 import dk.cachet.carp.clients.domain.data.AnyConnectedDeviceDataCollector
 import dk.cachet.carp.clients.domain.data.DataListener
 import dk.cachet.carp.clients.domain.data.DeviceDataCollectorFactory
